traders/requirements.txt
asgiref==3.7.2
Django==4.0.10
django-mongoengine==0.5.5
djongo==1.3.6
dnspython==2.4.2
mongoengine==0.27.0
packaging==23.2
plotly==5.18.0
pymongo==3.12.3
pytz==2023.3.post1
sqlparse==0.2.4
tenacity==8.2.3
django-crispy-forms
crispy-bootstrap4
djangorestframework
djangorestframework-simplejwt
Pillow
PyJWT